:recycle: refactor: Refatoração do uso de Route Handlers para Server Actions.
